Fix(install): if `$breaks` is equal to `$name`

+++ misc/scripts/install-local.sh
@@ -317,7 +317,7 @@ fi
 
 if echo -n "$depends" > /dev/null 2>&1; then
 	if [[ -n "$breaks" ]]; then
-		if dpkg-query -W -f='${Status}' "${breaks}" 2> /dev/null | grep "^install ok installed" > /dev/null 2>&1; then
+		if dpkg-query -W -f='${Status}' "${breaks}" 2> /dev/null | grep "^install ok installed" > /dev/null 2>&1 && ! pacstall -L | grep "${breaks}"; then
 			# Check if anything in breaks variable is installed already
 			fancy_message error "${RED}$name${NC} breaks $breaks, which is currently installed by apt"
 			error_log 13 "install $PACKAGE"
